Mini RAG System/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.10.3
langchain==0.3.11
langchain-community==0.3.10
faiss-cpu==1.9.0
pypdf==5.1.0
openai==1.58.1
langchain-openai==0.2.13
sentence-transformers==3.3.1
langchain-google-genai==2.0.6
python-multipart==0.0.20
